+++ python/spider-py/src/spider_py/type/utils.py
