--- python/src/spider/__init__.py
@@ -1,19 +1,0 @@
-"""Spider package root."""
-
-from spider.client import chain, Data, Driver, group, TaskContext, TaskGraph
-from spider.type import Double, Float, Int8, Int16, Int32, Int64
-
-__all__ = [
-    "Data",
-    "Double",
-    "Driver",
-    "Float",
-    "Int8",
-    "Int16",
-    "Int32",
-    "Int64",
-    "TaskContext",
-    "TaskGraph",
-    "chain",
-    "group",
-]
