--- python/src/spider/client/__init__.py
@@ -1,15 +1,0 @@
-"""Spider python client."""
-
-from .data import Data
-from .driver import Driver
-from .task import TaskContext
-from .taskgraph import chain, group, TaskGraph
-
-__all__ = [
-    "Data",
-    "Driver",
-    "TaskContext",
-    "TaskGraph",
-    "chain",
-    "group",
-]
